move fungible storage state
