+++ neo3/contracts/native/__init__.py
@@ -1,5 +1,5 @@
 from .nativecontract import NativeContract
-from .fungible import (FungibleToken, NeoToken, GasToken)
+from .fungible import (FungibleToken, NeoToken, GasToken, FungibleTokenStorageState)
 from .policy import PolicyContract
 from .oracle import OracleContract
 from .designate import DesignateRole, DesignationContract
