+++ neo3/storage/storageitem.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
-from typing import cast, Union, Type, Optional, Callable, TypeVar
+from typing import Type, Optional
 from enum import IntFlag
-from neo3 import vm
-from neo3.core import serialization, utils, IClonable, IInteroperable, Size as s
-from neo3.core.serialization import BinaryReader, BinaryWriter
+from neo3.core import serialization, utils, IClonable, Size as s
 
 
 class StorageFlags(IntFlag):
@@ -64,35 +62,32 @@ class StorageItem(serialization.ISerializable, IClonable):
         return cls(b'')
 
 
-class FungibleTokenStorageState(IInteroperable, serialization.ISerializable):
-    """
-    Helper class for NEP17 balance state
-
-    Use the from_storage() method if you're working with a DB snapshot and intend to modify the state values.
-    It will ensure that the cache is updated automatically.
-    """
-
-    def __init__(self):
-        super(FungibleTokenStorageState, self).__init__()
-        self.balance: vm.BigInteger = vm.BigInteger.zero()
-
-    def __len__(self):
-        return len(self.balance.to_array())
-
-    def serialize(self, writer: BinaryWriter) -> None:
-        writer.write_var_bytes(self.balance.to_array())
-
-    def deserialize(self, reader: BinaryReader) -> None:
-        self.balance = vm.BigInteger(reader.read_var_bytes())
-
-    def to_stack_item(self, reference_counter: vm.ReferenceCounter) -> vm.StackItem:
-        struct = vm.StructStackItem(reference_counter)
-        struct.append(vm.IntegerStackItem(self.balance))
-        return struct
-
-    @classmethod
-    def from_stack_item(cls, stack_item: vm.StackItem):
-        si = cast(vm.StructStackItem, stack_item)
-        c = cls()
-        c.balance = si[0].to_biginteger()
-        return c
+# class FungibleTokenStorageState(IInteroperable, serialization.ISerializable):
+#     """
+#     Helper class for NEP17 balance state
+#     """
+#
+#     def __init__(self):
+#         super(FungibleTokenStorageState, self).__init__()
+#         self.balance: vm.BigInteger = vm.BigInteger.zero()
+#
+#     def __len__(self):
+#         return len(self.balance.to_array())
+#
+#     def serialize(self, writer: BinaryWriter) -> None:
+#         writer.write_var_bytes(self.balance.to_array())
+#
+#     def deserialize(self, reader: BinaryReader) -> None:
+#         self.balance = vm.BigInteger(reader.read_var_bytes())
+#
+#     def to_stack_item(self, reference_counter: vm.ReferenceCounter) -> vm.StackItem:
+#         struct = vm.StructStackItem(reference_counter)
+#         struct.append(vm.IntegerStackItem(self.balance))
+#         return struct
+#
+#     @classmethod
+#     def from_stack_item(cls, stack_item: vm.StackItem):
+#         si = cast(vm.StructStackItem, stack_item)
+#         c = cls()
+#         c.balance = si[0].to_biginteger()
+#         return c
