+++ tests/contracts/interop/test_contract_interop.py
@@ -1,10 +1,8 @@
 import unittest
-import json
 from neo3 import vm, contracts, storage
 from neo3.network import payloads
-from neo3.contracts import syscall_name_to_int
 from neo3.core import to_script_hash, types, cryptography
-from tests.contracts.interop.utils import test_engine, test_block, test_tx
+from tests.contracts.interop.utils import test_engine, test_tx
 from copy import deepcopy
 
 """
